+++ src/openeo_gfmap/__init__.py
@@ -6,14 +6,14 @@ by leveraging the power of OpenEO (http://openeo.org/).
 More information available in the README.md file.
 """
 
-from .backend import Backend
+from .backend import _BackendType
 from .fetching import FetchType
 from .metadata import FakeMetadata
 from .spatial import BoundingBoxExtent, SpatialContext
 from .temporal import TemporalContext
 
 __all__ = [
-    "Backend",
+    "_BackendType",
     "SpatialContext",
     "BoundingBoxExtent",
     "TemporalContext",