+++ src/openeo_gfmap/fetching/s2.py
@@ -7,7 +7,7 @@ from typing import Callable
 import openeo
 from geojson import GeoJSON
 
-from openeo_gfmap.backend import Backend
+from openeo_gfmap.backend import _BackendType
 from openeo_gfmap.metadata import FakeMetadata
 from openeo_gfmap.spatial import BoundingBoxExtent, SpatialContext
 from openeo_gfmap.temporal import TemporalContext
@@ -196,30 +196,30 @@ def _get_s2_l2a_default_processor(
 
 
 SENTINEL2_L2A_BACKEND_MAP = {
-    Backend.TERRASCOPE: {
+    _BackendType.TERRASCOPE: {
         "fetch": partial(_get_s2_l2a_default_fetcher, collection_name="SENTINEL2_L2A"),
         "preprocessor": partial(
             _get_s2_l2a_default_processor, collection_name="SENTINEL2_L2A"
         ),
     },
-    Backend.CDSE: {
+    _BackendType.CDSE: {
         "fetch": partial(_get_s2_l2a_default_fetcher, collection_name="SENTINEL2_L2A"),
         "preprocessor": partial(
             _get_s2_l2a_default_processor, collection_name="SENTINEL2_L2A"
         ),
     },
-    Backend.CDSE_STAGING: {
-        "fetch": partial(_get_s2_l2a_default_fetcher, collection_name="SENTINEL2_L2A"),
-        "preprocessor": partial(
-            _get_s2_l2a_default_processor, collection_name="SENTINEL2_L2A"
-        ),
-    },
-    Backend.FED: {
-        "fetch": partial(_get_s2_l2a_default_fetcher, collection_name="SENTINEL2_L2A"),
-        "preprocessor": partial(
-            _get_s2_l2a_default_processor, collection_name="SENTINEL2_L2A"
-        ),
-    },
+    # Backend.CDSE_STAGING: {
+    #     "fetch": partial(_get_s2_l2a_default_fetcher, collection_name="SENTINEL2_L2A"),
+    #     "preprocessor": partial(
+    #         _get_s2_l2a_default_processor, collection_name="SENTINEL2_L2A"
+    #     ),
+    # },
+    # Backend.FED: {
+    #     "fetch": partial(_get_s2_l2a_default_fetcher, collection_name="SENTINEL2_L2A"),
+    #     "preprocessor": partial(
+    #         _get_s2_l2a_default_processor, collection_name="SENTINEL2_L2A"
+    #     ),
+    # },
 }