+++ tests/tests_integration/preprocessing/test_cloud_masking.py
@@ -1,9 +1,10 @@
 import os
 from pathlib import Path
 
+import openeo
 import pytest
 
-from openeo_gfmap.backend import Backend, get_connection
+from openeo_gfmap.backend import _BackendType
 from openeo_gfmap.fetching import FetchType, build_sentinel2_l2a_extractor
 from openeo_gfmap.preprocessing import (
     bap_masking,
@@ -15,7 +16,7 @@ from openeo_gfmap.spatial import BoundingBoxExtent
 from openeo_gfmap.temporal import TemporalContext
 from openeo_gfmap.utils import quintad_intervals
 
-backends = [Backend.CDSE]
+backends = [_BackendType.CDSE]
 
 # Few fields around Mol, Belgium
 spatial_extent = BoundingBoxExtent(
@@ -38,8 +39,8 @@ temporal_extent = TemporalContext(start_date="2022-11-01", end_date="2023-02-28"
 @pytest.mark.skipif(
     os.environ.get("SKIP_INTEGRATION_TESTS") == "1", reason="Skip integration tests"
 )
-def test_bap_score(backend: Backend):
-    connection = get_connection(backend=backend)
+def test_bap_score(backend: _BackendType):
+    connection = openeo.connect(backend.default_url).authenticate_oidc()
 
     # Additional parameters
     fetching_parameters = {"fetching_resolution": 10.0}
@@ -86,8 +87,8 @@ def test_bap_score(backend: Backend):
     os.environ.get("SKIP_INTEGRATION_TESTS") == "1", reason="Skip integration tests"
 )
 @pytest.mark.parametrize("backend", backends)
-def test_bap_masking(backend: Backend):
-    connection = get_connection(backend=backend)
+def test_bap_masking(backend: _BackendType):
+    connection = openeo.connect(backend.default_url).authenticate_oidc()
 
     # Additional parameters
     fetching_parameters = {"fetching_resolution": 10.0}
@@ -139,8 +140,8 @@ def test_bap_masking(backend: Backend):
     os.environ.get("SKIP_INTEGRATION_TESTS") == "1", reason="Skip integration tests"
 )
 @pytest.mark.parametrize("backend", backends)
-def test_bap_quintad(backend: Backend):
-    connection = get_connection(backend=backend)
+def test_bap_quintad(backend: _BackendType):
+    connection = openeo.connect(backend.default_url).authenticate_oidc()
 
     # Additional parameters
     fetching_parameters = {"fetching_resolution": 10.0}